+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-
+### Fixed
+- Fix #46: `instance list --output-options Tags` panics if no tags are set for an instance
 
 ## [0.0.17] - 2018-06-18